scripts/bioflow_obm.py
import os
import sys
import torch
import numpy as np
import logging
from typing import List, Union, Dict, Any, Optional
from tqdm import tqdm

# Add root to python path to allow imports from open_biomed
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

# Removed BioMedGPT references and added placeholders for open-source models
class BioFlowOBM:
    def __init__(self, config_path: str = "configs/model/opensource_model.yaml"):
        self.config_path = config_path
        self.model = None  # Placeholder for open-source model

    def initialize(self):
        # Placeholder for initializing open-source model
        pass

    def process_data(self, data):
        # Placeholder for processing data using open-source model
        pass

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Example usage block
if __name__ == "__main__":
    # Create valid dummy data/config for test if needed
    print("This script is a library. Import OBM to use.")
    print("Example:")
    print("from scripts.bioflow_obm import OBM")
    print("obm = OBM()")
    print("vec = obm.encode_text('Biology is complex')")
