requirements.txt
torch_geometric==2.5.2
transformers==4.38.0
rdkit==2023.09.5
sentencepiece==0.2.0
selfies==2.1.2
pandas==2.2.3
lightning_utilities==0.11.9
torchmetrics==1.5.2
numpy==1.26.4
absl-py==2.1.0
easydict==1.13
ratelimiter==1.2.0.post0
openai==1.64.0
uvicorn===0.32.1
fastapi===0.115.5
oss2==2.19.1
dotenv==1.0.1
protobuf==5.28.3
scanpy==1.10.3
qdrant-client>=1.7.0
streamlit>=1.28.0
plotly>=5.18.0
scikit-learn>=1.3.0
